# specification/docs/api-design/openapi.yaml
openapi: 3.0.0
info:
  title: サークル練習参加抽選システム API
  description: 練習日への投票・重み付き抽選・結果確認を提供する REST API
  version: 1.0.0
servers:
  - url: /api/v1
security:
  - bearerAuth: []
...
paths:
...
  /users/me:
    get:
      tags: [users]
      summary: 自分のプロフィール取得
      responses:
        '200':
          description: プロフィール
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
    put:
      tags: [users]
      summary: 自分のプロフィール更新
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdateRequest'
      responses:
        '204':
          description: 更新成功
        '400':
          $ref: '#/components/responses/BadRequest'
  /users:
    get:
      tags: [users]
      summary: 名簿一覧（代表のみ。男女全体）
      description: REQ-007.1 / 007.1.1。学年・マネージャー区分での絞り込み、名前・学籍番号での検索に対応
      parameters:
        - { name: grade, in: query, schema: { type: integer, enum: [1, 2, 3] } }
        - { name: is_manager, in: query, schema: { type: boolean } }
        - { name: gender, in: query, schema: { $ref: '#/components/schemas/Gender' } }
        - { name: q, in: query, description: 名前・学籍番号の部分一致, schema: { type: string } }
        - { name: page, in: query, schema: { type: integer, default: 1 } }
        - { name: per_page, in: query, schema: { type: integer, default: 50 } }
      responses:
        '200':
          description: 名簿
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RosterPage'
        '403':
          $ref: '#/components/responses/Forbidden'
  /users/export:
    get:
      tags: [users]
      summary: 名簿CSVエクスポート（代表のみ）
      description: REQ-007.1.2。/users と同じ絞り込みパラメータを受け付け、結果を CSV で返す
      parameters:
        - { name: grade, in: query, schema: { type: integer, enum: [1, 2, 3] } }
        - { name: is_manager, in: query, schema: { type: boolean } }
        - { name: gender, in: query, schema: { $ref: '#/components/schemas/Gender' } }
        - { name: q, in: query, schema: { type: string } }
      responses:
        '200':
          description: CSVファイル
          content:
            text/csv:
              schema: { type: string }
        '403':
          $ref: '#/components/responses/Forbidden'
  /users/{userId}/role:
    put:
      tags: [users]
      summary: 代表権限の付与・剥奪（代表のみ）
      parameters:
        - $ref: '#/components/parameters/UserId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoleUpdateRequest'
      responses:
        '204':
          description: 更新成功
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /users/{userId}:
    delete:
      tags: [users]
      summary: メンバー無効化（論理削除・代表のみ）
      parameters:
        - $ref: '#/components/parameters/UserId'
      responses:
        '204':
          description: 無効化成功
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /practice-months:
    get:
      tags: [practice-months]
      summary: 月別練習の一覧（自分の性別グループのみ）
      parameters:
        - { name: year_month, in: query, schema: { type: string, example: '2026-08' } }
      responses:
        '200':
          description: 月別練習一覧
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PracticeMonth'
    post:
      tags: [practice-months]
      summary: 月別練習単位の作成（代表のみ）
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PracticeMonthCreateRequest'
      responses:
        '201':
          description: 作成成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PracticeMonth'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          description: 同じ月×性別が既に存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /practice-months/{pmId}:
    get:
      tags: [practice-months]
      summary: 月別練習の詳細（練習日一覧含む）
      parameters:
        - $ref: '#/components/parameters/PmId'
      responses:
        '200':
          description: 詳細
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PracticeMonthDetail'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [practice-months]
      summary: 投票期間・枠比率等の更新（代表のみ）
      parameters:
        - $ref: '#/components/parameters/PmId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PracticeMonthUpdateRequest'
      responses:
        '204':
          description: 更新成功
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /practice-months/{pmId}/practices:
    post:
      tags: [practice-months]
      summary: 練習日の追加（代表のみ）
      parameters:
        - $ref: '#/components/parameters/PmId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PracticeCreateRequest'
      responses:
        '201':
          description: 作成成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Practice'
        '403':
          $ref: '#/components/responses/Forbidden'
  /practices/suggestions:
    get:
      tags: [practice-months]
      summary: よく使う練習の候補（履歴から）
      description: |
        担当性別の過去の登録実績から「場所 + 開始/終了時刻」の組み合わせを
        利用回数の多い順に返す。日程登録時の入力を省くために使う (REQ-003.6 / D-014)。

        定員は組み合わせのキーに含めない（施設の空き状況で毎月変わるため）。
        返す `capacity` は同じ組み合わせで最後に登録された値で、初期値として使う想定。
      responses:
        '200':
          description: 候補一覧（最大10件）
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PracticeSuggestion'
        '403':
          $ref: '#/components/responses/Forbidden'
  /practices/{practiceId}:
    put:
      tags: [practice-months]
      summary: 練習日の更新（代表のみ）
      parameters:
        - $ref: '#/components/parameters/PracticeId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PracticeCreateRequest'
      responses:
        '204':
          description: 更新成功
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [practice-months]
      summary: 練習日の削除（代表のみ。抽選実行後は要確認）
      parameters:
        - $ref: '#/components/parameters/PracticeId'
        - name: force
          in: query
          description: 抽選実行後の削除は true が必要（REQ-003.4 の警告）
          schema: { type: boolean, default: false }
      responses:
        '204':
          description: 削除成功
        '409':
          description: 抽選実行済み（force=true が必要）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /practice-months/{pmId}/votes/me:
    get:
      tags: [votes]
      summary: 自分の投票状況
      parameters:
        - $ref: '#/components/parameters/PmId'
      responses:
        '200':
          description: 投票状況
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VoteStatus'
    put:
      tags: [votes]
      summary: 投票の登録・変更（practice_id 配列で全置換）
      parameters:
        - $ref: '#/components/parameters/PmId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VoteUpdateRequest'
      responses:
        '204':
          description: 投票成功
        '409':
          description: 投票期間外（締切後の変更は不可）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /settings:
    get:
      tags: [settings]
      summary: 抽選設定の取得（代表のみ・担当性別）
      responses:
        '200':
          description: 設定
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LotterySettings'
        '403':
          $ref: '#/components/responses/Forbidden'
    put:
      tags: [settings]
      summary: 落選救済係数 α の更新（代表のみ）
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LotterySettingsUpdateRequest'
      responses:
        '204':
          description: 更新成功
        '403':
          $ref: '#/components/responses/Forbidden'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    UserId:
      name: userId
      in: path
      required: true
      schema: { type: string, example: usr_01H0000000000000000000000 }
    PmId:
      name: pmId
      in: path
      required: true
      schema: { type: string, example: pmn_01H0000000000000000000000 }
    PracticeId:
      name: practiceId
      in: path
      required: true
      schema: { type: string, example: prc_01H0000000000000000000000 }
  responses:
    BadRequest:
      description: 入力不正
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    Forbidden:
      description: 権限不足・性別スコープ違反
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: リソースが存在しない
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Gender:
      type: string
      enum: [male, female]
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code: { type: string, example: VALIDATION_ERROR }
            message: { type: string }
            details:
              type: array
              items:
                type: object
                properties:
                  field: { type: string }
                  reason: { type: string }
...
    UserProfile:
      type: object
      properties:
        id: { type: string, example: usr_01H0000000000000000000000 }
        email: { type: string, format: email }
        name: { type: string }
        address: { type: string }
        phone_number: { type: string }
        grade: { type: integer, enum: [1, 2, 3] }
        gender: { $ref: '#/components/schemas/Gender' }
        faculty_department: { type: string }
        student_number: { type: string }
        is_manager: { type: boolean }
        role: { type: string, enum: [member, representative] }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    UserUpdateRequest:
      type: object
      properties:
        name: { type: string }
        address: { type: string }
        phone_number: { type: string }
        grade: { type: integer, enum: [1, 2, 3] }
        faculty_department: { type: string }
        student_number: { type: string }
        is_manager: { type: boolean }
    RosterPage:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/UserProfile'
        total: { type: integer }
        page: { type: integer }
        per_page: { type: integer }
    RoleUpdateRequest:
      type: object
      required: [role]
      properties:
        role: { type: string, enum: [member, representative] }
    PracticeMonth:
      type: object
      properties:
        id: { type: string, example: pmn_01H0000000000000000000000 }
        year_month: { type: string, example: '2026-08' }
        gender: { $ref: '#/components/schemas/Gender' }
        vote_starts_at: { type: string, format: date-time }
        vote_ends_at: { type: string, format: date-time }
        grade2_ratio:
          type: number
          nullable: true
          description: D-015 で廃止（学年別枠は practices 側で持つ）。以前のデータのみ値を持つ
        status: { type: string, enum: [draft, voting, closed, drawn, published] }
        published_at: { type: string, format: date-time, nullable: true }
    PracticeMonthDetail:
      allOf:
        - $ref: '#/components/schemas/PracticeMonth'
        - type: object
          properties:
            practices:
              type: array
              items:
                $ref: '#/components/schemas/Practice'
    PracticeMonthCreateRequest:
      type: object
      required: [year_month, vote_starts_at, vote_ends_at]
      properties:
        year_month: { type: string, example: '2026-08' }
        vote_starts_at: { type: string, format: date-time }
        vote_ends_at: { type: string, format: date-time }
        practices:
          type: array
          description: 月の練習日をまとめて登録（REQ-003.1）
          items:
            $ref: '#/components/schemas/PracticeCreateRequest'
    PracticeMonthUpdateRequest:
      type: object
      properties:
        vote_starts_at: { type: string, format: date-time }
        vote_ends_at: { type: string, format: date-time }
        # grade2_ratio は D-015 で廃止。学年別枠は PUT /practice-months/{pmId}/quotas で設定する
    Practice:
      type: object
      properties:
        id: { type: string, example: prc_01H0000000000000000000000 }
        practice_date: { type: string, format: date }
        starts_at: { type: string, example: '18:00' }
        ends_at: { type: string, example: '21:00' }
        location: { type: string, example: 第一体育館 }
        capacity: { type: integer, example: 30 }
        vote_count: { type: integer, description: 現在の投票数（一覧表示用） }
    PracticeCreateRequest:
      type: object
      required: [practice_date, starts_at, ends_at, location, capacity]
      properties:
        practice_date: { type: string, format: date }
        starts_at:
          type: string
          pattern: '^\d{2}:(00|30)$'
          example: '18:00'
          description: 30分単位のみ (REQ-003.5 / D-014)
        ends_at:
          type: string
          pattern: '^\d{2}:(00|30)$'
          example: '21:30'
          description: 30分単位のみ (REQ-003.5 / D-014)
        location: { type: string }
        capacity: { type: integer, minimum: 1 }
    PracticeSuggestion:
      type: object
      description: 過去の登録実績から作る入力候補 (REQ-003.6 / D-014)
      properties:
        location: { type: string, example: 目白小 }
        starts_at: { type: string, example: '18:00' }
        ends_at: { type: string, example: '21:30' }
        capacity: { type: integer, example: 30, description: 同じ組み合わせで最後に登録された定員 }
        use_count: { type: integer, example: 4, description: 過去に登録された回数（多い順に並ぶ） }
    VoteStatus:
      type: object
      properties:
        practice_month_id: { type: string }
        voted_practice_ids:
          type: array
          items: { type: string }
        editable: { type: boolean, description: 締切前かどうか }
    VoteUpdateRequest:
      type: object
      required: [practice_ids]
      properties:
        practice_ids:
          type: array
          description: 投票する練習日IDの全量（全置換）
          items: { type: string }
...
    LotterySettings:
      type: object
      properties:
        gender: { $ref: '#/components/schemas/Gender' }
        rescue_alpha: { type: number, example: 0.2 }
    LotterySettingsUpdateRequest:
      type: object
      required: [rescue_alpha]
      properties:
        rescue_alpha: { type: number, minimum: 0 }